# Conflicts:
#	src/main/java/com/academy/services/EmailService.java
#	src/main/java/com/academy/services/MemberService.java
#	src/main/resources/sql-scripts/populateGlobalConfigurations.sql

+++ src/main/java/com/academy/services/ServiceProviderService.java
@@ -7,10 +7,12 @@ import com.academy.dtos.service_provider.ServiceProviderResponseDTO;
 import com.academy.exceptions.AuthenticationException;
 import com.academy.exceptions.EntityNotFoundException;
 import com.academy.exceptions.MemberNotFoundException;
+import com.academy.models.appointment.Appointment;
 import com.academy.models.member.Member;
 import com.academy.models.service.Service;
 import com.academy.models.service.service_provider.ProviderPermissionEnum;
 import com.academy.models.service.service_provider.ServiceProvider;
+import com.academy.repositories.AppointmentRepository;
 import com.academy.repositories.ServiceProviderRepository;
 import com.academy.utils.Utils;
 import jakarta.transaction.Transactional;
@@ -30,6 +32,7 @@ public class ServiceProviderService {
     private final ServiceService serviceService;
     private final AuthenticationFacade authenticationFacade;
     private final ProviderPermissionService providerPermissionService;
+    private final AppointmentRepository appointmentRepository;
 
     @Autowired
     public ServiceProviderService(ServiceProviderRepository serviceProviderRepository,
@@ -37,13 +40,14 @@ public class ServiceProviderService {
                                   ProviderPermissionService providerPermissionService,
                                   MemberService memberService,
                                   @Lazy ServiceService serviceService,
-                                  AuthenticationFacade authenticationFacade) {
+                                  AuthenticationFacade authenticationFacade, AppointmentRepository appointmentRepository) {
         this.serviceProviderRepository = serviceProviderRepository;
         this.serviceProviderMapper = serviceProviderMapper;
         this.memberService = memberService;
         this.serviceService = serviceService;
         this.providerPermissionService = providerPermissionService;
         this.authenticationFacade = authenticationFacade;
+        this.appointmentRepository = appointmentRepository;
     }
 
     public static void checkIfValidPermissions(List<ProviderPermissionEnum> newPermissions) throws BadRequestException {
@@ -53,6 +57,12 @@ public class ServiceProviderService {
 
     public List<ServiceProviderResponseDTO> getAllServiceProviders() {
         return serviceProviderRepository.findAll().stream()
+                .map(serviceProviderMapper::toResponseDTO)
+                .toList();
+    }
+
+    public List<ServiceProviderResponseDTO> getServiceProvidersByServiceId(Long serviceId) {
+        return serviceProviderRepository.findByServiceId(serviceId).stream()
                 .map(serviceProviderMapper::toResponseDTO)
                 .toList();
     }
@@ -74,6 +84,10 @@ public class ServiceProviderService {
         return serviceProviderRepository.findAllByProviderId(id);
     }
 
+    public List<ServiceProvider> getAllByServiceOwnerId(Long ownerId) {
+        return serviceProviderRepository.findAllByServiceOwnerId(ownerId);
+    }
+
     @Transactional
     public ServiceProviderResponseDTO createServiceProviderWithDTO(ServiceProviderRequestDTO dto) throws BadRequestException {
         ServiceProvider serviceProvider = createServiceProvider(dto);
@@ -91,6 +105,13 @@ public class ServiceProviderService {
 
         if(!checkIfHasPermissionToAddServiceProvider(loggedMember, service, dto.isServiceCreation()))
             throw new AuthenticationException("You do not have permission to create a Service Provider");
+
+        Optional<ServiceProvider> existingProvider = serviceProviderRepository.findByServiceIdAndProviderId(service.getId(), member.getId());
+        if(existingProvider.isPresent()) {
+            ServiceProvider serviceProvider = existingProvider.get();
+            existingProvider.get().setActive(true);
+            return serviceProviderRepository.save(serviceProvider);
+        }
 
         ServiceProvider serviceProvider = serviceProviderMapper.toEntity(dto);
         serviceProvider.setProvider(member);
@@ -139,9 +160,20 @@ public class ServiceProviderService {
     }
 
     @Transactional
-    public void deleteServiceProvider(long id) {
+    public void deleteServiceProvider(long id) throws BadRequestException {
         ServiceProvider serviceProvider = serviceProviderRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException(ServiceProvider.class, id));
+
+        String loggedUsername = authenticationFacade.getUsername();
+        Member loggedMember = memberService.getMemberByUsername(loggedUsername);
+        Service service = serviceProvider.getService();
+
+        if(!checkIfHasPermissionToAddServiceProvider(loggedMember, service, false))
+            throw new AuthenticationException("You do not have permission to delete a Service Provider");
+
+        if (serviceProvider.getPermissions().stream().anyMatch(permission -> permission.getPermission().equals(ProviderPermissionEnum.OWNER) ) ) {
+            throw new BadRequestException("Cannot remove owner from service.");
+        }
 
         providerPermissionService.deletePermissionsFromServiceProvider(serviceProvider);
         serviceProvider.setActive(false);
@@ -223,4 +255,16 @@ public class ServiceProviderService {
         return serviceProviderRepository.findProvidersByServiceIdAndPermission(serviceId, permission);
     }
 
+    public void updateRating(Long id){
+        Double rating = appointmentRepository.findAverageRatingByServiceProvider_Id(id);
+        System.out.println("Updating rating for service provider with id " + id + " to " + rating);
+        if (rating != null) {
+            int roundedRating = Math.toIntExact(Math.round(rating));
+            ServiceProvider provider = serviceProviderRepository.findById(id)
+                    .orElseThrow(() -> new EntityNotFoundException(ServiceProvider.class, id));
+            provider.setRating(roundedRating);
+            serviceProviderRepository.save(provider);
+        }
+
+    }
 }
